refactor: extract processing utilities
